requirements.txt
cx_Oracle
flask
flask-cors
python-dotenv